requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==2.0.12
commonmark==0.9.1
dataclasses==0.8
idna==3.7
packaging==21.3
Pygments==2.14.0
pyparsing==3.1.2
python-dotenv==0.20.0
pyzabbix==1.3.1
requests==2.27.1
rich==12.6.0
selenium==3.141.0
soupsieve==2.3.2.post1
typing-extensions==4.1.1
urllib3==1.26.19
